Consolidated util and Utility folders
* Added DeformDataParser.hpp to CMake sources to allow IDEs like Visual
Studio to see it
