+++ SimulationIOFilters/Utility/SourceList.cmake
@@ -3,6 +3,7 @@ set(${PLUGIN_NAME}_UTILITY_HDRS
   ${${PLUGIN_NAME}_SOURCE_DIR}/${PLUGIN_NAME}Filters/Utility/BsamFileWriter.h
   ${${PLUGIN_NAME}_SOURCE_DIR}/${PLUGIN_NAME}Filters/Utility/OnScaleTableFileWriter.h
   ${${PLUGIN_NAME}_SOURCE_DIR}/${PLUGIN_NAME}Filters/Utility/EntriesHelper.h
+  ${${PLUGIN_NAME}_SOURCE_DIR}/${PLUGIN_NAME}Filters/Utility/DeformDataParser.hpp
 )
 
 set(${PLUGIN_NAME}_UTILITY_SRCS
